boot/bfs: Fix compilation errors in Stream.h and BPlusTree.h
This commit addresses two compilation issues in the BFS boot loader implementation:

1. `src/system/boot/loader/file_systems/bfs/Stream.h`:
   - The `Stream` class defined a member function `Node()` which shadowed the global `Node` type.
   - This caused `Node *NodeFactory(...)` to fail as `Node` was interpreted as the function.
   - Fixed by changing the return type to `::Node *`.

2. `src/add-ons/kernel/file_systems/bfs/BPlusTree.h`:
   - The definition of `FragmentAt` used `int8 index` while the declaration used `int32 index`.
   - This caused a "no declaration matches" error.
   - Fixed by changing the definition to use `int32 index`.

+++ src/add-ons/kernel/file_systems/bfs/BPlusTree.h
@@ -604,7 +604,7 @@ bplustree_node::IsLeaf() const
 
 
 inline duplicate_array*
-bplustree_node::FragmentAt(int8 index) const
+bplustree_node::FragmentAt(int32 index) const
 {
 	return (duplicate_array*)((off_t*)this + index * (NUM_FRAGMENT_VALUES + 1));
 }
